+++ src/components/Modal/Modal.stories.tsx
@@ -1,73 +1,91 @@
 ﻿import * as React from "react";
+import { Meta } from "@storybook/react";
 import { Modal, ModalBody, ModalFooter, ModalSize } from "./Modal";
 import { CustomButton, ButtonType } from "../CustomButton/CustomButton";
-import * as bootstrap from "bootstrap";
 
 export default {
     title: "Components/Modal",
     component: Modal,
-};
+} as Meta;
 
-const OpenModalButton = ({ modalId, text }: { modalId: string; text: string }) => {
-    const openModal = () => {
-        const modalElement = document.getElementById(modalId);
-        if (modalElement) {
-            const bsModal = new bootstrap.Modal(modalElement);
-            bsModal.show();
-        }
-    };
+const ModalExample = ({
+    modalId,
+    modalSize,
+    scrollable = false,
+    preventCloseOnOutsideClick = true,
+    content = "Innehåll...",
+}: {
+    modalId: string;
+    modalSize: ModalSize;
+    scrollable?: boolean;
+    preventCloseOnOutsideClick?: boolean;
+    content?: React.ReactNode;
+}) => {
+    const [isOpen, setIsOpen] = React.useState(false);
 
-    return <CustomButton buttonText={text} buttonType={ButtonType.general} onClick={openModal} />;
+    const handleOpen = () => setIsOpen(true);
+    const handleClose = () => setIsOpen(false);
+
+    return (
+        <>
+            <CustomButton
+                buttonText="Öppna modal"
+                buttonType={ButtonType.general}
+                onClick={handleOpen}
+            />
+            <Modal
+                modalId={modalId}
+                header="Exempelmodal"
+                modalSize={modalSize}
+                scrollable={scrollable}
+                preventCloseOnOutsideClick={preventCloseOnOutsideClick}
+                isOpen={isOpen}
+                onClose={handleClose}
+            >
+                <ModalBody>{content}</ModalBody>
+                <ModalFooter>
+                    <CustomButton
+                        buttonText="Stäng"
+                        buttonType={ButtonType.general}
+                        data-bs-dismiss="modal"
+                    />
+                </ModalFooter>
+            </Modal>
+        </>
+    );
 };
 
 export const ModalSmall = () => (
-    <div>
-        <OpenModalButton modalId="modalSmall" text="Open Small Modal" />
-        <Modal header="Small Modal" id="modalSmall" modalSize={ModalSize.small}>
-            <ModalBody>Content...</ModalBody>
-            <ModalFooter>Footer</ModalFooter>
-        </Modal>
-    </div>
+    <ModalExample modalId="modal-small" modalSize={ModalSize.small} />
 );
 
 export const ModalNormalPreventCloseOnOutsideClickFalse = () => (
-    <div>
-        <OpenModalButton modalId="modalNormalPreventCloseOnOutsideClickFalse" text="Open Normal Modal (can close outside)" />
-        <Modal header="Normal Modal" id="modalNormalPreventCloseOnOutsideClickFalse" modalSize={ModalSize.normal} preventCloseOnOutsideClick={false}>
-            <ModalBody>Click outside modal to close</ModalBody>
-            <ModalFooter>Footer</ModalFooter>
-        </Modal>
-    </div>
+    <ModalExample
+        modalId="modal-normal"
+        modalSize={ModalSize.normal}
+        preventCloseOnOutsideClick={false}
+    />
 );
 
 export const ModalLarge = () => (
-    <div>
-        <OpenModalButton modalId="modalLarge" text="Open Large Modal" />
-        <Modal header="Large Modal" id="modalLarge" modalSize={ModalSize.large}>
-            <ModalBody>Content...</ModalBody>
-            <ModalFooter>Footer Content</ModalFooter>
-        </Modal>
-    </div>
+    <ModalExample modalId="modal-large" modalSize={ModalSize.large} />
 );
 
 export const ModalFullscreen = () => (
-    <div>
-        <OpenModalButton modalId="modalFullscreen" text="Open Fullscreen Modal" />
-        <Modal header="Fullscreen Modal" id="modalFullscreen" modalSize={ModalSize.full}>
-            <ModalBody>Fullscreen content...</ModalBody>
-            <ModalFooter>Footer Content</ModalFooter>
-        </Modal>
-    </div>
+    <ModalExample modalId="modal-fullscreen" modalSize={ModalSize.full} />
 );
 
 export const ModalScrollable = () => (
-    <div>
-        <OpenModalButton modalId="modalScrollable" text="Open Scrollable Modal" />
-        <Modal header="Scrollable Modal" id="modalScrollable" modalSize={ModalSize.normal} scrollable={true}>
-            <ModalBody>
-                {Array(25).fill(<p>Rader...</p>)}
-            </ModalBody>
-            <ModalFooter>Footer Content</ModalFooter>
-        </Modal>
-    </div>
+    <ModalExample
+        modalId="modal-scrollable"
+        modalSize={ModalSize.normal}
+        scrollable={true}
+        content={
+            <>
+                {Array.from({ length: 40 }, (_, i) => (
+                    <p key={i}>Rader {i + 1}</p>
+                ))}
+            </>
+        }
+    />
 );